feat: enhance Next.js configuration with image optimization and bundle analyzer support
- Added deviceSizes, imageSizes, and formats to optimize image loading.
- Included framer-motion in optimizePackageImports for better performance.
- Implemented conditional bundle analyzer integration based on ANALYZE environment variable.
- Updated package.json to add an analyze script for easier bundle analysis.
