+++ frontend/app/politician/[id]/PoliticianPageClient.tsx
@@ -1,6 +1,7 @@
 "use client"
 
 import { useEffect, useRef, useState } from "react"
+import NextImage from "next/image"
 import { AlertTriangle, BarChart2, CalendarCheck, HelpCircle, MessageSquare, Trophy, UserPlus } from "lucide-react"
 import { useRouter } from "next/navigation"
 import { CitationLink } from "@/components/CitationLink"
@@ -76,7 +77,7 @@ function Section({
             }}
         >
             <div className="flex items-center gap-3 mb-4">
-                <img src={icon} alt={title} className="w-6 h-6 object-contain" />
+                <NextImage src={icon} alt={title} width={24} height={24} className="w-6 h-6 object-contain" />
                 <Text variant="h4" weight="bold" className="text-gray-900 dark:text-white">
                     {title}
                 </Text>
@@ -310,21 +311,21 @@ function ContactSection({ politician }: { politician: Politician }) {
             <div className="grid gap-3">
                 {contact?.email && (
                     <div className="flex items-center gap-2">
-                        <img src="/logo/location.png" alt="Email" className="w-4 h-4" />
+                        <NextImage src="/logo/location.png" alt="Email" width={16} height={16} className="w-4 h-4" />
                         <Text variant="body" className="text-gray-700 dark:text-gray-300">{contact.email}</Text>
                         <CitationLink citation={contact_citations?.email} label="Email" />
                     </div>
                 )}
                 {contact?.phone && (
                     <div className="flex items-center gap-2">
-                        <img src="/logo/location.png" alt="Phone" className="w-4 h-4" />
+                        <NextImage src="/logo/location.png" alt="Phone" width={16} height={16} className="w-4 h-4" />
                         <Text variant="body" className="text-gray-700 dark:text-gray-300">{contact.phone}</Text>
                         <CitationLink citation={contact_citations?.phone} label="Phone" />
                     </div>
                 )}
                 {contact?.address && (
                     <div className="flex items-center gap-2">
-                        <img src="/logo/location.png" alt="Address" className="w-4 h-4" />
+                        <NextImage src="/logo/location.png" alt="Address" width={16} height={16} className="w-4 h-4" />
                         <Text variant="body" className="text-gray-700 dark:text-gray-300">{contact.address}</Text>
                         <CitationLink citation={contact_citations?.address} label="Address" />
                     </div>
@@ -380,7 +381,7 @@ function ContactSection({ politician }: { politician: Politician }) {
                             <span className="inline-flex items-center gap-1">
                                 <a href={social_media.website} target="_blank" rel="noopener noreferrer"
                                     className="text-green-600 dark:text-green-400 hover:underline text-sm flex items-center gap-1">
-                                    <img src="/logo/location.png" alt="Website" className="w-3 h-3" />
+                                    <NextImage src="/logo/location.png" alt="Website" width={12} height={12} className="w-3 h-3" />
                                     Website
                                 </a>
                                 <CitationLink citation={social_media_citations?.website} label="Website" />
@@ -706,12 +707,13 @@ export default function PoliticianPageClient({
                                     width={128}
                                     height={128}
                                     priority
+                                    quality={75}
                                     sizes="128px"
                                     className="w-32 h-32 rounded-2xl object-cover border-4 border-orange-200 dark:border-orange-800"
                                 />
                             ) : (
                                 <div className="w-32 h-32 rounded-2xl bg-linear-to-br from-orange-100 to-orange-200 dark:from-orange-900/40 dark:to-orange-800/40 flex items-center justify-center flex-shrink-0 border-4 border-orange-200 dark:border-orange-800">
-                                    <img src="/logo/Parliament.png" alt="Politician" className="w-16 h-16 object-contain" />
+                                    <NextImage src="/logo/Parliament.png" alt="Politician" width={64} height={64} className="w-16 h-16 object-contain" />
                                 </div>
                             )}
                         </div>
@@ -726,19 +728,19 @@ export default function PoliticianPageClient({
 
                             <div className="space-y-1.5">
                                 <div className="flex items-center gap-2 hero-info-1">
-                                    <img src="/logo/Parliament.png" alt="Party" className="w-4 h-4 object-contain opacity-60" />
+                                    <NextImage src="/logo/Parliament.png" alt="Party" width={16} height={16} className="w-4 h-4 object-contain opacity-60" />
                                     <Text variant="body" className="text-gray-700 dark:text-gray-300">
                                         <span className="font-semibold">Party:</span> {party}
                                     </Text>
                                 </div>
                                 <div className="flex items-center gap-2 hero-info-2">
-                                    <img src="/logo/Assembly.png" alt="Constituency" className="w-4 h-4 object-contain opacity-60" />
+                                    <NextImage src="/logo/Assembly.png" alt="Constituency" width={16} height={16} className="w-4 h-4 object-contain opacity-60" />
                                     <Text variant="body" className="text-gray-700 dark:text-gray-300">
                                         <span className="font-semibold">Constituency:</span> {p.constituency}
                                     </Text>
                                 </div>
                                 <div className="flex items-center gap-2 hero-info-3">
-                                    <img src="/logo/location.png" alt="State" className="w-4 h-4 object-contain opacity-60" />
+                                    <NextImage src="/logo/location.png" alt="State" width={16} height={16} className="w-4 h-4 object-contain opacity-60" />
                                     <Text variant="body" className="text-gray-700 dark:text-gray-300">
                                         <span className="font-semibold">State:</span> {p.state}
                                     </Text>
