+++ frontend/components/marketing/index.ts
@@ -1,0 +1,4 @@
+export { default as HeroSection } from "./HeroSection";
+export { default as FeaturesSection } from "./FeaturesSection";
+export { default as ContributeSection } from "./ContributeSection";
+export { default as ContributorsSection } from "./ContributorsSection";
